requirements.txt
requests==2.32.4
python-dotenv==1.1.1
selenium==4.33.0